requirements.txt
cx_Freeze==6.15.16
eventlet==0.36.1
inkex==1.0.1
lxml==4.8.0
mediapipe==0.10.9
microrep==1.0.0
numpy==1.24.2
numpy==1.21.5
opencv_contrib_python==4.9.0.80
Pillow==9.4.0
Pillow==10.3.0
PySide6==6.6.1
PySide6==6.6.3.1
PySide6_Addons==6.6.1
PySide6_Essentials==6.6.1
python-socketio==5.11.2
qroundprogressbar==1.0.0
shiboken6==6.6.1
svgutils==0.3.4
